@@ -1,23 +1,16 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import { ReactDOM } from "react-dom/client";
-import { BrowserRouter, Route, Routes } from "react-router-dom"
-import Header from "./components/header/Header";
-import Footer from "./components/footer/Footer";
-import "./app.css";
-import Contacto from "./components/pages/Contacto/Contacto";
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import "./index.css";
+import {BrowserRouter } from 'react-router-dom'
+import Header from "./components/Header";
+import Footer from "./components/Footer"
 
-ReactDOM.render(
-  <React.StrictMode>
-  <Header />
-  <BrowserRouter>
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/Contact" element={<Contacto />} />
-    </Routes>
-  </BrowserRouter>
-  <Footer />
-  </React.StrictMode>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+<BrowserRouter>
+    <Header />
+    <App />
+    <Footer />
+</BrowserRouter>
 );
